fix: 🐛 provide consistency of `define` between rax & react (#236)
* fix: 🐛 provide consistency of `define` between rax & react

✅ Closes: 235
