+++ packages/build-plugin-component/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.8.1
+
+- [feat] support babelPlugins in `type: rax`.
+- [fix] provide consistency of `define` between rax & react ([#235](https://github.com/ice-lab/iceworks-cli/issues/235))
+
 ## 1.8.0
 
 - [feat] use `disableGenerateStyle` to bypass `style.js` ([#4705](https://github.com/alibaba/ice/issues/4705))
