+++ packages/build-plugin-component/src/configs/rax/userConfig.js
@@ -100,4 +100,8 @@ module.exports = [
     name: WECHAT_MINIPROGRAM,
     validation: 'object',
   },
+  {
+    name: 'babelPlugins',
+    validation: 'array',
+  },
 ];
