+++ packages/build-plugin-component/src/configs/react/userConfig.js
@@ -65,10 +65,6 @@ module.exports = [
     validation: 'object',
   },
   {
-    name: 'define',
-    validation: 'object',
-  },
-  {
     name: 'disableGenerateStyle',
     validation: 'boolean',
   },
