+++ packages/build-plugin-component/src/utils/getUMDWebpack.js
@@ -2,6 +2,7 @@ const path = require('path');
 const { upperFirst, camelCase } = require('lodash');
 const { getWebpackConfig } = require('build-scripts-config');
 const { defaultDynamicImportLibraries } = require('../compiler/depAnalyze');
+const setDefine = require('../useConfig/define');
 
 module.exports = ({ context, compileOptions, extNames, hasMain }) => {
   const mode = 'production';
@@ -146,19 +147,7 @@ module.exports = ({ context, compileOptions, extNames, hasMain }) => {
   }
 
   if (define) {
-    const defineVariables = {};
-    Object.keys(define).forEach((defineKey) => {
-      defineVariables[defineKey] = JSON.stringify(define[defineKey]);
-    });
-
-    if (config.plugins.get('DefinePlugin')) {
-      config
-        .plugin('DefinePlugin')
-        .tap((args) => [Object.assign(...args, defineVariables)]);
-    } else {
-      config.plugin('DefinePlugin')
-        .use(webpack.DefinePlugin, [defineVariables]);
-    }
+    setDefine(config, define, context);
   }
 
   return config;
